serial_core: Provide a transceiver interface
Before this change, to enable transceiver control for RS-485 ports,
client drivers had to implement the TIOCSRS485/TIOCGRS485 ioctls. This
change provides a common transceiver interface so that we have a common
way for all client drivers to do transceiver control.

When registering a new port, client drivers will have the option of
providing a struct txvr_ops which has callbacks for transceiver
operations. This makes it easier to have a common UART driver (like the
8250) for both RS-232 and RS-485 ports and transceiver operations can
be added for the RS-485 ports alone.

Signed-off-by: Jaeden Amero <jaeden.amero@ni.com>
Signed-off-by: Xander Huff <xander.huff@ni.com>
Acked-by: Karthik Manamcheri <karthik.manamcheri@ni.com>
Acked-by: Ben Shelton <ben.shelton@ni.com>
Natinst-ReviewBoard-ID: 88262

8250: Allow client drivers to control transceivers

This change allows 8250 client drivers to register txvr_ops, so that
they can implement transceiver control for RS-485 ports.

Signed-off-by: Jaeden Amero <jaeden.amero@ni.com>
Signed-off-by: Xander Huff <xander.huff@ni.com>
Acked-by: Karthik Manamcheri <karthik.manamcheri@ni.com>
Acked-by: Ben Shelton <ben.shelton@ni.com>
Natinst-ReviewBoard-ID: 88262

8250: Add a driver for NI 16550 UART

The National Instruments (NI) 16550 is a standard 16550 with larger
FIFOs and embedded RS-485 transceiver control circuitry. This patch
adds a driver that can operate this UART.

Signed-off-by: Jaeden Amero <jaeden.amero@ni.com>
Acked-by: Karthik Manamcheri <karthik.manamcheri@ni.com>
[gratian: fix conflict with aa42db44f670 ("8250: of: remove remnants of generic of_serial driver")]
Signed-off-by: Gratian Crisan <gratian.crisan@ni.com>
[bstreiff: renumber out-of-tree definition of PORT_NI16550 from 118 to 123.]
Signed-off-by: Brandon Streiff <brandon.streiff@ni.com>
[gratian: renumber out-of-tree definition of PORT_NI16550 due to upstream addition 9e8d5470325f
 ("serial: sunplus-uart: Add Sunplus SoC UART Driver")]
Signed-off-by: Gratian Crisan <gratian.crisan@ni.com>

8250_pnp: Make NIC7750 ID spawn a NI 16550 port

This commit adds a new ACPI ID, NIC7750, which will be used to represent
a National Instruments NI 16550 UART. To enable this driver, you need
to set CONFIG_SERIAL_8250_NI16550 in the kernel config.

Signed-off-by: Karthik Manamcheri <karthik.manamcheri@ni.com>
Acked-by: Jaeden Amero <jaeden.amero@ni.com>

8250_pnp: NI ACPI IDs start with NIC

To determine if a device is a NI device, we should check the first
three characters of the ACPI ID (All NI IDs start with "NIC").

Signed-off-by: Karthik Manamcheri <karthik.manamcheri@ni.com>
Acked-by: Jaeden Amero <jaeden.amero@ni.com>

8250_pnp: Support dual-mode UART through PMR

The NI CVS-1459RT Controller has a NI 16550 UART which can be connected
to either a RS-232 transceiver or a RS-485 transceiver. This dual-mode
support is implemented through the Port Mode Register (PMR). This
change will add support for the 8250 PNP driver to read the PMR and
based on the mode, will register the appropriate driver.

Signed-off-by: Karthik Manamcheri <karthik.manamcheri@ni.com>
Acked-by: Jaeden Amero <jaeden.amero@ni.com>
Acked-by: Justin Tang <justin.tang@ni.com>

8250: NI 16550 can have different FIFO sizes

The NI 16550 is a CPLD UART and depending on restrictions on CPLD size,
different implementations can have different FIFO sizes. This change
provides two NI 16550 types, one with a smaller 16-byte FIFOs and the
other with the original 128-byte FIFOs.

Signed-off-by: Karthik Manamcheri <karthik.manamcheri@ni.com>
Acked-by: Justin Tang <justin.tang@ni.com>
Acked-by: Jaeden Amero <jaeden.amero@ni.com>
[bstreiff: renumber PORT_NI16550_*]
Signed-off-by: Brandon Streiff <brandon.streiff@ni.com>
[gratian: renumber PORT_NI16550_*]
Signed-off-by: Gratian Crisan <gratian.crisan@ni.com>

8250_pnp: Provide NI_16BYTE_FIFO flag

This change will provide a flag, NI_16BYTE_FIFO, which when used as the
flag for ACPI IDs will spawn a NI UART with a 16-byte FIFO (instead
of the default 128-byte FIFO)

Signed-off-by: Karthik Manamcheri <karthik.manamcheri@ni.com>
Acked-by: Justin Tang <justin.tang@ni.com>

8250_pnp: Make NIC7772 spawn a dual-mode UART

This commit adds a new ACPI ID, NIC7772. This ID represents a
National Instruments NI 16550 UART which is dual-mode capable (Port
Mode Register is implemented) and has 16-byte FIFOs. To use this port,
you need to set CONFIG_SERIAL_8250_NI16550 in the kernel config.

Signed-off-by: Karthik Manamcheri <karthik.manamcheri@ni.com>

ACPI / PNP: Add IDs for NI 16550 to list

Add the two PNP IDs for NI 16550 serial ports to the PNP device ID
list, so they are enumerated as PNP devices.

Signed-off-by: Nathan Sullivan <nathan.sullivan@ni.com>
Signed-off-by: Brad Mouring <brad.mouring@ni.com>
Acked-by: Xander Huff <xander.huff@ni.com>

Natinst-ReviewBoard-ID: 135498
Natinst-CAR-ID: 584586

8250_pnp: Add new ID NIC792B for NI 16550 UART 25MHz

Add the new ACPI ID NIC792B into 8250_pnp driver to support
NI 16550 UART with 25MHz clock.

Signed-off-by: Stephen Lee <slee@ni.com>
Signed-off-by: Brad Mouring <brad.mouring@ni.com>
Acked-by: Hui Chun Ong <hui.chun.ong@ni.com>
Reviewed-by: Jason Smith <jason.smith@ni.com>
Natinst-ReviewBoard-ID: 166881

serial: 8250_ni16550: Add support for NIC7A69

This device uses a 33.333MHz clock with a prescaler value of 1.125 to
achieve baud rates within 1% error up to 921600 baud. We query ACPI
standard _DSD properties to determine if each port is an RS-232 or RS-485
port.

Signed-off-by: Kyle Roeschley <kyle.roeschley@ni.com>
Natinst-ReviewBoard-ID: 286745

8250: Use proper set_mctrl callback for ni16550 clock divider

Commit b4a29b94804c ("serial: 8250: Move Alpha-specific quirk out of the
core") removed the mcr_mask and mcr_force members from struct
uart_8250_port. Upstream they are only used for an Alpha-specific
quirk. The Alpha code was instead refactored to use the proper
'->set_mctrl' callback.

The out-of-tree patches for the 8250 ni16550 serial driver rely on
mcr_force to set UART_MCR_CLKSEL (divide clock by 4) for certain clock
rate configurations. Refactor the code to use a '->set_mctrl' callback
instead to set this MCR bit flag.

Fixes: ff4f5de15f3b ("8250_pnp: Add new ID NIC792B for NI 16550 UART 25MHz")
Fixes: c170227b9b1f ("serial: 8250_ni16550: Add support for NIC7A69")
Signed-off-by: Gratian Crisan <gratian.crisan@ni.com>