nirtfeatures: restructure init function to close holes
In nirtfeatures_acpi_add, we create several sysfs files. As currently
implemented, there is a window where access to a sysfs file may cause
our spinlock to be used before it's been initialized, or may cause us
to write an incorrect value to an I/O port. We can close this window
by moving the creation of the sysfs files closer to the end of the
function.

Signed-off-by: Jeff Westfahl <jeff.westfahl@ni.com>
[cvadrevu: fixed trivial conflict due to dropping 9cbd2f25fe01
("nirtfeatures: set HARD_BOOT_N if necessary when the driver loads")]
Signed-off-by: Chaitanya Vadrevu <chaitanya.vadrevu@emerson.com>