time: Make the clocksource watchdog user configurable
The clocksource watchdog is used to detect instabilities in the current
clocksource. This is a beneficial feature on new/unknown hardware however it
can create problems by falsely triggering when the watchdog wraps. The
reason is that an interrupt storm and/or high priority (FIFO/RR) tasks can
preempt the timer softirq long enough for the watchdog to wrap if it has a
limited number of bits available by comparison with the main clocksource.

One observed example is on a Intel Baytrail platform where TSC is the main
clocksource, HPET is disabled due to a hardware bug and acpi_pm gets
selected as the watchdog clocksource.

Provide the option to disable the clocksource watchdog for hardware where
the clocksource stability has been validated.

Signed-off-by: Gratian Crisan <gratian.crisan@ni.com>
[gratian: fix trivial conflict with fc153c1c58cb ("clocksource: Add a Kconfig option for WATCHDOG_MAX_SKEW")]
Signed-off-by: Gratian Crisan <gratian.crisan@ni.com>
[gratian: fix trivial conflict with 7cf8f44a5a1c ("x86: fs: kmsan: disable CONFIG_DCACHE_WORD_ACCESS")]
Signed-off-by: Gratian Crisan <gratian.crisan@ni.com>