ath6kl: Retry SDIO function initialization on failure
Sometimes the AR6234 gives us a error "ath6kl: Unable to enable
sdio func: -62".  This indicates that the enable bit for the
function did not result in the function's status bit in the
SDIO_CCCR_IORx register getting set within a reasonable period
of time (like 400ms).  The guess at this point is that the radio
is coming up in some strange state and the power-on reset did
not clean things up as it should have.  This occurs about once
every thousand or so reboots.

Experimentation has shown that resetting the device after this
has happened allows the function enable to correctly set the
function's bit in the SDIO_CCCR_IORx.  This change will reset
the AR6234 and try the function enable again if the module
parameter boot_attempts is set to >0.  This has been tested over
several thousand reboots, and the driver will recover correctly
when it would have otherwise failed to initialize.

Signed-off-by: James Minor <james.minor@ni.com>
Signed-off-by: Brad Mouring <brad.mouring@ni.com>
Acked-by: Nathan Sullivan <nathan.sullivan@ni.com>
Acked-by: Xander Huff <xander.huff@ni.com>

Natinst-ReviewBoard-ID: 135864
Natinst-CAR-ID: 579000