Revert "time: Always make sure wall_to_monotonic isn't positive"
This reverts commit e1d7ba8735551ed79c7a0463a042353574b96da3.

NI has a use case that involves distributed networked Linux devices
that need to share the same concept of time using a mechanism like
IEEE-1588 or 802.1AS.  The “master” device (ie. the device with the
time all other devices will be synchronized to) is often a device
that boots up set to the Posix Epoch, mainly because it lacks a
battery-backed RTC. (note: the existence of an RTC does not prevent
a device from booting up at or very near the Posix Epoch, it just
greatly reduces the likelihood).

If a slave device attempted to synchronize its CLOCK_REALTIME to
that of the master – and the master’s time was < Epoch+slave uptime,
that slave would not be able to synchronize.

This use case is believed to be very common among embedded devices,
especially those without RTCs.

Long term:
We (NI Timing & Sync) are planning on engaging with the upstream
community to educate them on our use case and hopefully put a
different solution in place which solves the original problem
(preventing a negative boot time representation) while also allowing
our use case to continue working as it did prior to the change we’re
reverting. Once that happens, we can drop this revert.

Signed-off-by: Brad Mouring <brad.mouring@ni.com>
Reported-by: Vineeth Acharya <vineeth.acharya@ni.com>
Tested-by: Rick Ratzel <rick.ratzel@ni.com>
Natinst-CAR-ID: 629499
[bstreiff: reduced control flow in do_settimeofday64 due to unassigned 'ret']
Signed-off-by: Brandon Streiff <brandon.streiff@ni.com>
[gratian: fix conflict with b8ac29b40183 ("timekeeping: contribute wall clock to rng on time change")]
Signed-off-by: Gratian Crisan <gratian.crisan@ni.com>