nati_x86_64_defconfig: Add mousedev module
Before the 4.14 upgrade we were compiling in the mousedev support.
After the upgrade it is not selected by default presumably because it
is superseded by the input event interface.

Re-add the mousedev interface driver as a module. It provides mouse
access via /dev/input/mouseX and /dev/input/mice. The majority of
newer devices work well with the generic input device event interface
but there is still a fair amount of legacy hardware out there that
might require this module.

Signed-off-by: Gratian Crisan <gratian.crisan@ni.com>
Acked-by: Julia Cartwright <julia.cartwright@ni.com>
Natinst-ReviewBoard-ID: 253941