Revert "softirq: Check preemption after reenabling interrupts" for irq_poll
irq_poll_cpu_dead() defers its softirq processing until a random point.
This is true for RT & !RT.
The checks in irq_poll_softirq() are not needed because it is invoked
from within the softirq handler.
The check in irq_poll_complete() is not needed because there is no task
wake here.
The check in irq_poll_sched() should not be needed. Should as in all
handler should invoke it from their interrupt service routing. I'm not sure
about lpfc and mpt3sas due to long call chain but if it is invoked from a
preemptible context then it is delayed to random point in time with and without
PREEMPT_RT.

Signed-off-by: Sebastian Andrzej Siewior <bigeasy@linutronix.de>