net: Update the seqcount_t removal from Qdisc.
This is an all-in-one patch udpating the seqcount_t removal from Qdisc
to the version recently posted on the list.
   Try to simplify the gnet_stats and remove qdisc->running sequence counter.
   https://lore.kernel.org/all/20211016084910.4029084-1-bigeasy@linutronix.de/

Signed-off-by: Sebastian Andrzej Siewior <bigeasy@linutronix.de>

net: Update the seqcount_t removal from Qdisc.

This is an all-in-one patch updating the series based on what has been
merged upstream plus the individual fixes:

   net/sched: sch_ets: properly init all active DRR list handles
   net: sched: Allow statistics reads from softirq.
   net: sched: fix logic error in qdisc_run_begin()
   net: sched: remove one pair of atomic operations
   net: stats: Read the statistics in ___gnet_stats_copy_basic() instead of adding.

Signed-off-by: Sebastian Andrzej Siewior <bigeasy@linutronix.de>

net: Update the seqcount_t removal from Qdisc.

This is an all-in-one patch udpating the seqcount_t removal from Qdisc
to the version recently posted on the list.
   Try to simplify the gnet_stats and remove qdisc->running sequence counter.
   https://lore.kernel.org/all/20211016084910.4029084-1-bigeasy@linutronix.de/

Signed-off-by: Sebastian Andrzej Siewior <bigeasy@linutronix.de>