nati_x86_64_defconfig: readd CONFIG_SERIAL_8250_RUNTIME_UARTS=0
When we rebased the defconfig options from 4.14 to 5.6 (and then
further to 5.10), the option CONFIG_SERIAL_8250_RUNTIME_UARTS=0 was
accidentally dropped.

Without this, with the default option of 4, the 8250 driver will
automatically add four hardcoded serial ports at the legacy PC serial
port addresses (from the SERIAL_PORT_DFNS array), regardless of whether
or not they're present in the ACPI tables. On sbRIO platforms, only
the first serial port is at 0x3F8-- the remaining ones are at different
locations, so with the default setting we get a ttyS1-ttyS3 that are
at invalid addresses and shift the remaining eight serial ports (which
do come from ACPI) down to new indices, potentially breaking userspace.

This commit restores the behavior as shipped with 4.14.

Natinst-AZDO-ID: 1722341
Fixes: 05b901147db7 ("nati_x86_64_defconfig: update to 5.6.14-rt7")
Signed-off-by: Brenda Streiff <brenda.streiff@ni.com>