printk: Check only for migration in printk_deferred_*().
Atomic context is not required by the implementation. The only
requirement is that the caller does not migrate to another CPU between
the _enter() and _exit() invocation. The reason is to increment and
decrement the per-CPU variable on the same CPU.
Checking for migration only allows to use deferred printk on PREEMPT_RT
when only sleeping locks are acquired.

Check for disabled migration instead for atomic context in
printk_deferred_*()

Signed-off-by: Sebastian Andrzej Siewior <bigeasy@linutronix.de>